+++ R/estimate_gps.R
@@ -9,28 +9,35 @@
 #' @param Y A vector of observed outcome variable.
 #' @param w A vector of observed continuous exposure variable.
 #' @param c A data frame of observed covariates variable.
-#' @param pred_model The selected prediction model.
 #' @param gps_model Model type which is used for estimating GPS value, including
 #' parametric (default) and non-parametric.
 #' @param internal_use If TRUE will return helper vectors as well. Otherwise,
-#'  will return original data + GPS value.
-#' @param params Includes list of params that is used internally. Unrelated
+#'  will return original data + GPS values.
+#' @param params Includes list of parameters that are used internally. Unrelated
 #'  parameters will be ignored.
-#' @param nthread An integer value that represents then number threads to use by
-#'  internal packages.
+#' @param sl_lib A vector of prediction algorithms.
+#' @param nthread An integer value that represents the number threads to be used
+#' in a shared memory system.
 #' @param ...  Additional arguments passed to the model.
 #'
 #' @return
-#' The function returns a list of 6 objects according to the following order:
-#'   - Original data set + GPS, counter, row_index values (Y, w, GPS, counter,
-#'    row_index, c)
+#' The function returns a S3 object. Including the following:
+#'   - Original data set + GPS, counter, row_index values (Y, w, GPS,
+#'   counter_weight, row_index, c)
 #'   - e_gps_pred
 #'   - e_gps_std_pred
 #'   - w_resid
 #'   - gps_mx (min and max of gps)
 #'   - w_mx (min and max of w).
+#'   - used_params
+#'
+#' @note
 #' If \code{internal.use} is set to be FALSE, only original data set + GPS will
 #' be returned.
+#'
+#' The outcome variable is not used in estimating the GPS value. However, it is
+#' used in compiling the data set with GPS values.
+#'
 #'
 #' @export
 #'
@@ -39,7 +46,6 @@
 #' data_with_gps <- estimate_gps(m_d$Y,
 #'                               m_d$treat,
 #'                               m_d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
-#'                               pred_model = "sl",
 #'                               gps_model = "parametric",
 #'                               internal_use = FALSE,
 #'                               params = list(xgb_max_depth = c(3,4,5),
@@ -51,18 +57,17 @@
 estimate_gps <- function(Y,
                          w,
                          c,
-                         pred_model,
                          gps_model = "parametric",
                          internal_use = TRUE,
                          params = list(),
+                         sl_lib = c("m_xgboost"),
                          nthread = 1,
                          ...){
 
-  sl_lib = NULL
   start_time <- proc.time()
 
-  # Check passed arguments
-  check_args_estimate_gps(pred_model, gps_model, ...)
+  # Check passed arguments -----------------------------------------------------
+  check_args_estimate_gps(gps_model, ...)
 
   dot_args <- list(...)
   arg_names <- names(dot_args)
@@ -71,12 +76,25 @@ estimate_gps <- function(Y,
     assign(i,unlist(dot_args[i],use.names = FALSE))
   }
 
-  # Generate SL wrapper library for each type of prediction algorithms
+  # Check if data has missing value(s) -----------------------------------------
+  if (sum(is.na(w)) > 0){
+    logger::log_warn("Vector w has {sum(is.na(w))} missing values.")
+  }
+
+  if (sum(is.na(c)) > 0){
+    logger::log_warn(
+      "Confounders data.frame (c) has {sum(is.na(c))} missing values.")
+  }
+
+  # Generate SL wrapper library for each type of prediction algorithms ---------
   sl_lib_internal = NULL
+  used_params <- list()
   for (item in sl_lib){
-    wrapper_generated <- gen_wrap_sl_lib(lib_name = item, params, nthread = nthread)
-    if (wrapper_generated){
+    wrapper_generated_res <- gen_wrap_sl_lib(lib_name = item, params,
+                                             nthread = nthread)
+    if (wrapper_generated_res[[1]]){
       sl_lib_internal <- c(sl_lib_internal,paste(item,"_internal", sep=""))
+      used_params <- c(used_params, wrapper_generated_res[[2]])
     } else {
       sl_lib_internal <- c(sl_lib_internal, item)
     }
@@ -84,7 +102,7 @@ estimate_gps <- function(Y,
 
   if (gps_model == "parametric"){
 
-    e_gps <- train_it(target = w, input = c, pred_model,
+    e_gps <- train_it(target = w, input = c,
                       sl_lib_internal = sl_lib_internal, ...)
     e_gps_pred <- e_gps$SL.predict
     e_gps_std_pred <- stats::sd(w - e_gps_pred)
@@ -93,10 +111,10 @@ estimate_gps <- function(Y,
 
   } else if (gps_model == "non-parametric"){
 
-    e_gps <- train_it(target = w, input = c, pred_model,
+    e_gps <- train_it(target = w, input = c,
                       sl_lib_internal = sl_lib_internal, ...)
     e_gps_pred <- e_gps$SL.predict
-    e_gps_std <- train_it(target = abs(w-e_gps_pred), input = c, pred_model,
+    e_gps_std <- train_it(target = abs(w-e_gps_pred), input = c,
                            sl_lib_internal = sl_lib_internal, ...)
     e_gps_std_pred <- e_gps_std$SL.predict
     w_resid <- compute_resid(w,e_gps_pred,e_gps_std_pred)
@@ -111,9 +129,9 @@ estimate_gps <- function(Y,
 
   w_mx <- compute_min_max(w)
   gps_mx <- compute_min_max(gps)
-  counter <- (w*0)+0 # initialize counter.
+  counter_weight <- (w*0)+0 # initialize counter.
   row_index <- seq(1,length(w),1) # initialize row index.
-  dataset <- cbind(Y,w,gps,counter, row_index, c)
+  dataset <- cbind(Y,w,gps,counter_weight,row_index, c)
 
   # Logging for debugging purposes
   logger::log_debug("Min Max of treatment: {paste(w_mx, collapse = ', ')}")
@@ -138,9 +156,19 @@ estimate_gps <- function(Y,
   logger::log_debug("Wall clock time to run estimate_gps function: ",
                     " {(end_time - start_time)[[3]]} seconds.")
 
+
+  result <- list()
+  class(result) <- "cgps_gps"
+  result$dataset <- dataset
+  result$used_params <- used_params
+
   if (internal_use){
-    return(list(dataset, e_gps_pred, e_gps_std_pred, w_resid, gps_mx, w_mx))
-  } else {
-    return(list(dataset))
+    result$e_gps_pred <- e_gps_pred
+    result$e_gps_std_pred <- e_gps_std_pred
+    result$w_resid <- w_resid
+    result$gps_mx <- gps_mx
+    result$w_mx <- w_mx
   }
+
+  invisible(result)
 }