+++ NEWS.md
@@ -1,13 +1,45 @@
 ## CausalGPS (development)
 
 ### Fixed
-* 
+* `estimate_npmetric_erf` assigns user-defined log file.
 
 ### Changed
-* Vectorized population compilation.
+
+* The entire data set is trimmed based on trimming quantiles.
+* `earth` and `ranger` are not installed automatically. They can be installed manually if needed.
+* `sysdata.rda` is modified to reflect transition from `counter` and `ipw` to `counter_weight`
+* `counter_weight` is used as a counter or weight, in `matching` or `weighting`
+approaches. `counter` and `ipw` are dropped.
+* `sl_lib` becomes a required argument.
+* The package has been transferred into NSAPH-Software Github account.
+* Summary function of `gpsm_pspop` S3 object returns details of the adjusting process. 
+
 
 ### Added
-*
+* `pseodo_pop` also includes covariate column names.
+
+### Removed 
+
+* `pred_model` argument dropped. The package only predicts using SuperLearner.
+
+
+## CausalGPS 0.2.8 (2022-06-22)
+
+### Fixed
+* Message for not implemented methods changed to reduce misunderstanding.
+* Empty counter will raise error in estimating non-parameteric response function. 
+
+### Changed
+* matching_l1 returns frequency table instead of entire vector. 
+* Vectorized population compilation and used data.table for multi-thread assignment.
+* Removed nested parallelism in compiling pseudo population, which results in close control on memory.
+* estimate_npmetric_erf also returns optimal h and risk values.
+
+### Added
+* `estimate_gps` returns the optimal hyperparameters.
+* `estimate_gps` returns S3 object. 
+* Internal xgboost approach support `verbose` parameter.
+* Pseudo-population object now report the parameters that are used for the best covariate balance.
 
 
 ## CausalGPS 0.2.7 (2022-02-04)
@@ -28,7 +60,7 @@
 ## CausalGPS 0.2.6 (2021-09-06)
 
 ### Added
-* added the status of optimized compile to generate_psuedo_pop function output.
+* added the status of optimized compile to generate_pseudo_pop function output.
 * compute_closest_wgps accepts the number of user-defined threads.
 
 ### Changed
@@ -116,7 +148,7 @@
 
 ### Added
 
-*  User defined bin sequence in compiling speudo population.
+*  User defined bin sequence in compiling pseudo population.
 *  Non-parametric option for estimating GPS.
 *  Adaptive approach to transform features in training sessions.
 *  Cpp code for computing pair of w and GPS.